LSH object
