+++ schemadiscovery/src/main/scala/Main.scala
@@ -49,11 +49,10 @@ object Main {
       // Sort and print all distinct patterns found
       DataToPattern.printSortedPatterns(allPatterns)
 
-      val dataForLSH = prepareDataForLSH(allPatterns, spark)
-      val model = setupLSH(dataForLSH)
+      val dataForLSH = LSH.prepareDataForLSH(allPatterns, spark)
+      val model = LSH.setupLSH(dataForLSH)
       val hashedData = model.transform(dataForLSH)
       hashedData.show(false)
-
 
     } finally {
       spark.stop()
@@ -73,41 +72,6 @@ object Main {
       .csv(filePath)
   }
 
-  def setupLSH(data: DataFrame): BucketedRandomProjectionLSHModel = {
-    val brp = new BucketedRandomProjectionLSH()
-      .setBucketLength(2.0)
-      .setNumHashTables(3)
-      .setInputCol("features")
-      .setOutputCol("hashes")
 
-    brp.fit(data)
-  }
-
-  def prepareDataForLSH(allPatterns: mutable.Map[Seq[String], mutable.ListBuffer[Seq[Any]]], spark: SparkSession): DataFrame = {
-    import spark.implicits._
-
-    // Convert the map to a DataFrame with an array of Doubles
-    val data = allPatterns.flatMap { case (_, listBuffer) =>
-      listBuffer.map { row =>
-        row.map {
-          case num: Number => num.doubleValue()
-          case other => other.hashCode.toDouble
-        }.toArray
-      }
-    }.toSeq.toDF("array")
-
-    // Find the maximum length of any vector
-    val maxLength = data.select(max(size($"array"))).as[Int].collect()(0)
-
-    // UDF to pad vectors to the maximum length
-    val toPaddedVector = udf((array: Seq[Double]) => {
-      Vectors.dense((array ++ Array.fill(maxLength - array.length)(0.0)).toArray)
-    })
-
-    // Convert arrays to vectors and pad them to ensure consistent length
-    val dataForLSH = data.withColumn("features", toPaddedVector($"array")).drop("array")
-
-    dataForLSH
-  }
 
 }
